Feat: 모델 출력. 메모리 누수 발생중. 해결 필요.

+++ Engine/Private/UIObject.cpp
@@ -179,7 +179,7 @@ bool CUIObject::Consume_MouseRightClick()
 
 bool CUIObject::Check_MouseOver(POINT fPos)
 {
-	return static_cast<CRect_Transform*>( m_pTransformCom)->Is_InRect(_float2(fPos.x, fPos.y));
+	return static_cast<CRect_Transform*>( m_pTransformCom)->Is_InRect(_float2((float)fPos.x, (float)fPos.y));
 }
 
 CUIObject* CUIObject::Find_FocusedUI(POINT fPos)
